RepeatedTaskQueue test helper
When using a RepeatedTaskQueue with an explicit release delay queue in a
test, tests must remember to automatically release all pending actions
when the queue is shutdown. This change adds a
RepeatedTaskQueue.forTesting companion method that returns a
RepeatedTaskQueue with this behavior built in.
